+++ src/barr_commands/barr_cmd_fo.h
@@ -1,0 +1,12 @@
+#ifndef BARR_CMD_FO_H_
+#define BARR_CMD_FO_H_
+
+#include "barr_defs.h"
+
+#define BARR_FO_DIR ".barr/fo"
+
+#define BARR_FO_REPORT_FILE ".barr/fo/report"
+
+barr_i32 BARR_command_fo(barr_i32 argc, char **argv);
+
+#endif  // BARR_CMD_FO_H_